Move scripts around
