+++ src/lib/disemvoweler.svelte
@@ -1,0 +1,14 @@
+<script>
+	import { disemvowelWord, disemvowelBodyOfText } from '../scripts/disemvowel.js';
+</script>
+
+<div class="disemvoweler-container">
+	<h2>TODO: Disemvoweler</h2>
+</div>
+
+<style>
+	.disemvoweler-container {
+		background-color: beige;
+		padding: 20px;
+	}
+</style>
