+++ src/scripts/disemvowel.js
@@ -1,8 +1,9 @@
-function isVowel(letter) {
-    return ['a', 'e', 'i', 'o', 'u'].indexOf(letter.toLowerCase()) !== -1;
-}
+export function disemvowelWord(word) {
 
-function disemvowelWord(word) {
+    function isVowel(letter) {
+        return ['a', 'e', 'i', 'o', 'u'].indexOf(letter.toLowerCase()) !== -1;
+    }
+
     const firstLetter = word.charAt(0);
     const finalLetter = word.charAt(word.length - 1);
     const wordNoVowelsOrDoubles = word
@@ -35,7 +36,7 @@ function disemvowelWord(word) {
     else return wordNoVowelsOrDoubles;
 }
 
-function disemvowelBodyOfText(text) {
+export function disemvowelBodyOfText(text) {
     const stringArray = text.split(' ');
     const disemvoweledStringArray = stringArray.map((word) => disemvowelWord(word));
     const disemvoweledText = disemvoweledStringArray.join(' ');
