+++ src/scripts/turnSvgsIntoOutlineObjects.js
